Move OpInstruction to operations module, add basic design for Op
