+++ src/qutip_qip/circuit/__init__.py
@@ -4,7 +4,6 @@ from .instruction import (
     CircuitInstruction,
     GateInstruction,
     MeasurementInstruction,
-    OpInstruction,
 )
 from .simulator import CircuitResult, CircuitSimulator
 from .circuit import QubitCircuit
@@ -16,5 +15,4 @@ __all__ = [
     "CircuitInstruction",
     "GateInstruction",
     "MeasurementInstruction",
-    "OpInstruction",
 ]
